pyproject.toml
[project]
name = "Stream-Health-Monitor"
version = "0.0.2"
description = "Monitor video stream health from appliance to web"
authors = [{ name = "Dom Capparelli", email = "web@capparelli.ie" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.11,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aiohttp>=3.11.18",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "netifaces>=0.11.0",
    "prometheus-client>=0.21.1",
    "psutil>=7.0.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "pyvimeo>=1.1.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]

[project.urls]
Homepage = "https://capp3.github.io/Stream-Health-Monitor/"
Repository = "https://github.com/capp3/Stream-Health-Monitor"
Documentation = "https://capp3.github.io/Stream-Health-Monitor/"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-asyncio>=0.26.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.23.0",
    "mypy>=1.15.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "black>=25.1.0",
    "flake8>=7.2.0",
    "basedpyright>=1.29.0",
    "ipython>=9.1.0",
    "pyinstrument>=5.0.1",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"



[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]

